+++ plugins/weighted-r1-k1/actions/sendUserOperationWithSignatures.ts
@@ -13,11 +13,8 @@ import {
 import type { SmartAccount } from "permissionless/accounts"
 import { sendUserOperation as sendUserOperationBundler } from "permissionless/actions"
 import { prepareUserOperationRequest } from "permissionless/actions/smartAccount"
-import type { SendUserOperationParameters } from "permissionless/actions/smartAccount/sendUserOperation"
-import type {
-    EntryPoint,
-    GetEntryPointVersion
-} from "permissionless/types/entrypoint"
+import type { SendUserOperationParameters } from "permissionless/actions/smartAccount"
+import type { EntryPoint, GetEntryPointVersion } from "permissionless/types"
 import {
     type Chain,
     type Client,
@@ -32,11 +29,15 @@ import { encodeSignatures } from "../utils.js"
 
 export type SendUserOperationWithSignaturesParameters<
     entryPoint extends EntryPoint,
-    TAccount extends SmartAccount<entryPoint> | undefined =
-        | SmartAccount<entryPoint>
+    TTransport extends Transport = Transport,
+    TChain extends Chain | undefined = Chain | undefined,
+    TAccount extends
+        | SmartAccount<entryPoint, string, TTransport, TChain>
+        | undefined =
+        | SmartAccount<entryPoint, string, TTransport, TChain>
         | undefined
 > = Prettify<
-    SendUserOperationParameters<entryPoint, TAccount> & {
+    SendUserOperationParameters<entryPoint, TTransport, TChain, TAccount> & {
         signatures: Hex[]
     }
 >
@@ -45,13 +46,20 @@ export async function sendUserOperationWithSignatures<
     entryPoint extends EntryPoint,
     TTransport extends Transport = Transport,
     TChain extends Chain | undefined = Chain | undefined,
-    TAccount extends SmartAccount<entryPoint> | undefined =
-        | SmartAccount<entryPoint>
+    TAccount extends
+        | SmartAccount<entryPoint, string, TTransport, TChain>
+        | undefined =
+        | SmartAccount<entryPoint, string, TTransport, TChain>
         | undefined
 >(
     client: Client<TTransport, TChain, TAccount>,
     args: Prettify<
-        SendUserOperationWithSignaturesParameters<entryPoint, TAccount>
+        SendUserOperationWithSignaturesParameters<
+            entryPoint,
+            TTransport,
+            TChain,
+            TAccount
+        >
     >
 ): Promise<Hash> {
     const { account: account_ = client.account } = args