+++ plugins/permission/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/permissions
 
+## 5.4.6
+
+### Patch Changes
+
+- Patch for latest permissionless 0.1.45 support
+
 ## 5.4.5
 
 ### Patch Changes