+++ plugins/weighted-r1-k1/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/weighted-validator
 
+## 5.3.1
+
+### Patch Changes
+
+- Patch for latest permissionless 0.1.45 support
+
 ## 5.3.0
 
 ### Patch Changes